Move GenSDK & related functions to `packages`
These functions are used in multiple commands (install, package add, convert). Move them to their own package so `...cmd` packages don’t depend on each other.

There’s some more tidying and refactoring that we can do here, but this first PR only moves `project_sdk_link{,_test}.go`.
