+++ pkg/cmd/pulumi/packagecmd/package_gen_sdk.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2024, Pulumi Corporation.
+// Copyright 2016-2025, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import (
 	"github.com/spf13/cobra"
 
 	cmdCmd "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
+	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/packages"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/env"
@@ -63,7 +64,7 @@ If a folder either the plugin binary must match the folder name (e.g. 'aws' and 
 			}()
 
 			parameters := &plugin.ParameterizeArgs{Args: args[1:]}
-			pkg, _, err := SchemaFromSchemaSource(pctx, source, parameters,
+			pkg, _, err := packages.SchemaFromSchemaSource(pctx, source, parameters,
 				cmdCmd.NewDefaultRegistry(cmd.Context(), pkgWorkspace.Instance, nil, cmdutil.Diag(), env.Global()))
 			if err != nil {
 				return err
@@ -88,7 +89,7 @@ If a folder either the plugin binary must match the folder name (e.g. 'aws' and 
 
 			if language == "all" {
 				for _, lang := range []string{"dotnet", "go", "java", "nodejs", "python"} {
-					err := GenSDK(lang, out, pkg, overlays, local)
+					err := packages.GenSDK(lang, out, pkg, overlays, local)
 					if err != nil {
 						return err
 					}
@@ -96,7 +97,7 @@ If a folder either the plugin binary must match the folder name (e.g. 'aws' and 
 				fmt.Fprintf(os.Stderr, "SDKs have been written to %s\n", out)
 				return nil
 			}
-			err = GenSDK(language, out, pkg, overlays, local)
+			err = packages.GenSDK(language, out, pkg, overlays, local)
 			if err != nil {
 				return err
 			}
