+++ pkg/cmd/pulumi/packagecmd/package_info.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 
 	cmdCmd "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
+	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/packages"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/schema"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
@@ -61,7 +62,7 @@ The <provider> argument can be specified in the same way as in 'pulumi package a
 			}()
 
 			parameters := &plugin.ParameterizeArgs{Args: args[1:]}
-			pkg, _, err := SchemaFromSchemaSource(pctx, args[0], parameters,
+			pkg, _, err := packages.SchemaFromSchemaSource(pctx, args[0], parameters,
 				cmdCmd.NewDefaultRegistry(cmd.Context(), pkgWorkspace.Instance, nil, cmdutil.Diag(), env.Global()))
 			if err != nil {
 				return err
